+++ include/NPU/Dialect/NPU/IR/NPUDialect.h
@@ -1,0 +1,16 @@
+//===- NPUDialect.h - The npu dialect ---------------------------*- C++ -*-===//
+//
+// SPDX-FileCopyrightText: 2026 Olajide Badejo <olajideayomidebadejo@gmail.com>
+//
+// SPDX-License-Identifier: MIT
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef NPU_DIALECT_NPU_IR_NPUDIALECT_H
+#define NPU_DIALECT_NPU_IR_NPUDIALECT_H
+
+#include "mlir/IR/Dialect.h"
+
+#include "NPU/Dialect/NPU/IR/NPUOpsDialect.h.inc"
+
+#endif // NPU_DIALECT_NPU_IR_NPUDIALECT_H